% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BudgetIV_scalar.R
\name{BudgetIV_scalar}
\alias{BudgetIV_scalar}
\title{BudgetIV for single causal effect parameters}
\usage{
BudgetIV_scalar(
  beta_y,
  beta_phi,
  tau_vec,
  b_vec,
  delta_beta_y = NA,
  bounds_only = TRUE
)
}
\arguments{
\item{beta_y}{A \eqn{1 \times d_{Z}} vector representing the (estimated) 
cross covariance \eqn{\mathrm{Cov}(Y, Z)}.}

\item{beta_phi}{A \eqn{1 \times d_{Z}} vector representing the (estimated) 
cross covariance \eqn{\mathrm{Cov}(\Phi (X), Z)}.}

\item{tau_vec, }{A \eqn{1 \times K} vector of strictly increasing, positive budget thresholds. \eqn{K} is the number of budget groups.}

\item{b_vec}{A \eqn{1 \times K} vector of increasing positive integer budgets. 
Represents the constraint that at least \eqn{b_i} different values of \eqn{j}, 
the candidate instrument \eqn{Z_j} satisfies \eqn{\mathrm{Cov} (Y - \theta \Phi (X), Z_j) \leq \tau_j}.}

\item{delta_beta_y}{Either (a) a \eqn{1 \times d_{Z}} vector of positive half-widths for box-shaped confidence bounds on \code{beta_y};
or (b) the empty value \code{NA} for partial identification or feasible region estimates without uncertainty quantification.}

\item{bounds_only}{If TRUE (default), the output consists only of disjoint bounds. Otherwise, if FALSE, the output consists of bounds for 
possibly touching intervals (but never overlapping), as well as the budget assignment corresponding to each bound.}
}
\value{
A list of two entries: \code{intervals}, which is a two-column matrix with rows corresponding to disjoint bounds containing plausible values of \eqn{\theta}; 
and \code{points}, which is a one-column matrix consisting of lone plausible values of \eqn{\theta}---relevant when using \eqn{\tau_1 = 0}.
}
\description{
Partial identification and coverage of a causal effect parameter using summary statistics and budget constraint assumptions.
}
\details{
Instrumental variables are defined by three structural assumptions: (A1) they are associated with the treatment; 
(A2) they are unconfounded with the outcome; and (A3) they exclusively effect the outcome through the treatment. 
Assumption (A1) has a simple statistical test, whereas for many data generating processes (A2) and (A3) are 
unprovably false. 
The \code{BudgetIV} and \code{BudgetIV_scalar} algorithms allow for valid causal inference when some proportion, 
possibly a small minority, of candidate instruments satisfy both (A2) and (A3).

\code{BudgetIV} & \code{BudgetIV_scalar} assume a homogeneous treatment effect, which implies the separable structural 
equation \eqn{Y = \theta \Phi(X) + g_y(Z, \epsilon_x)}. 
The difference between the algorithms is that \code{BudgetIV_scalar} assumes \eqn{\Phi(X)} and \eqn{\theta} take
scalar values, which is exploited for super-exponential computational speedup and allows for causal inference
with thousands of candidate instruments \eqn{Z}.
Both methods assume ground truth knowledge of the functional form of \eqn{\Phi (X)}, e.g., a linear, 
logistic, Cox hazard, principal component based or other model. 
The parameter \eqn{\theta} captures the unknown treatment effect.
Violation of (A2) and/or (A3) will bias classical IV approaches through the statistical dependence
between \eqn{Z} and \eqn{g_y(Z, \epsilon_x)}, summarized by the covariance parameter 
\eqn{\gamma := \mathrm{Cov} (g_y(Z, \epsilon_x), Z)}.

\code{BudgetIV} & \code{BudgetIV_scalar} constrain \eqn{\gamma} through a series of positive thresholds 
\eqn{0 \leq \tau_1 < \tau_2 < \ldots < \tau_K} and corresponding integer budgets \eqn{0 < b_1 < b_2 < \ldots < b_K \leq d_Z}. 
It is assumed for each \eqn{i \in \{ 1, \ldots, K\}} that no more than \eqn{b_i} components of \eqn{\gamma} are greater in 
magnitude than \eqn{\tau_i}.
For instance, taking \eqn{d_Z = 100}, \eqn{K = 1}, \eqn{b_1 = 5} and \eqn{\tau_1 = 0} means 
assuming \eqn{5\%} of the \eqn{100} candidates are valid instrumental variables (in the sense that their ratio 
estimates \eqn{\theta_j := \mathrm{Cov}(Y, Z_j)/\mathrm{Cov}(\Phi(X), Z_j)} are unbiased).

With \code{delta_beta_y = NA}, \code{BudgetIV} & \code{BudgetIV_scalar} return the identified set
of causal effects that agree with both the budget constraints described above and the values of
\eqn{\mathrm{Cov}(Y, Z)} and \eqn{\mathrm{Cov}(Y, Z)}, assumed to be exactly precise. 
Unlike classical partial identification methods (see Manski (1990) ofr a canonical example), the non-convex mixed-integer
budget constraints yield a possibly disconnected identified set. 
Each connected subset has a different interpretation as to which of the candidate instruments \eqn{Z} 
are valid up to each threshold.
\code{BudgetIV_scalar} returns these interpretations alongside the corresponding bounds on \eqn{\theta}. 

When \code{delta_beta_y} is not null, it is used as box-constraints to quantify uncertainty in \code{beta_y}. 
In the examples, \code{delta_beta_y} is calculated through a Bonferroni correction and gives an (asymptotically) 
valid confidence set over \code{beta_y}. 
Under the so-called "no measurement error" (NOME) assumption (see Bowden et al. (2016)) which is commonly applied in Mendelian randomisation, it is
assumed that the estimate of \code{beta_y} is the dominant source of finite-sample uncertainty, with uncertainty in \code{beta_x}
entirely negligible. 
With an (asymptotically) valid confidence set over \code{delta_beta_y} and under the "no measurement error" assumption, \code{BudgetIV_scalar} 
returns an (asymptotically) valid confidence set for \eqn{\theta}.
}
\examples{
 
set.seed(42)

# Experiment to test .




}
\references{
Jordan Penn, Lee Gunderson, Gecia Bravo-Hermsdorff,
Ricardo Silva, and David Watson. (2024). BudgetIV: Optimal Partial Identification of Causal Effects with Mostly Invalid Instruments. \emph{arXiv}
preprint, 2411.06913.

Jack Bowden, Fabiola Del Greco M, Cosetta Minelli, George Davey Smith, Nuala A Sheehan, and John R Thompson. (2016). Assessing the suitability of summary data for 
two-sample Mendelian randomization analyses using MR-Egger regression: the role of the I^2 statistic. \emph{Int. J. Epidemiol.} 46.6, pp. 1985--1998.

Charles F Manski. (1990). Nonparametric bounds on treatment effects. \emph{Am. Econ. Rev.} 80.2, pp. 219--323.
}
